* commit '719e2c478ed2a20d74890f25b9f503c355d69b03':
  Added conversion back to RobotSide.
  Fixed a test.
  Fixed up the ROS API.
  Updated scripts.
  Trying to trick bamboo to not fail because of weird timeout declaration.
  Added statics.
  Added test that enforce the declaration of static values for enums.
  Renamed the byte fields of all packets to follow the new convention.
  Added test to assert byte field naming convention.
  Replaced enum fields with byte fields in all packets.
  Changed QueueableMessage to have 1 setter/getter per field.
  Nuked ValkyrieLowLevelControlModeMessage.
  Added test for enum fields in Packets.

+++ ihmc-avatar-interfaces/src/test/java/us/ihmc/avatar/roughTerrainWalking/AvatarFootstepDataMessageSwingTrajectoryTest.java
@@ -83,7 +83,7 @@ public abstract class AvatarFootstepDataMessageSwingTrajectoryTest implements Mu
       double transferTime = robotModel.getWalkingControllerParameters().getDefaultTransferTime();
       double initialTransferTime = robotModel.getWalkingControllerParameters().getDefaultInitialTransferTime();
       FootstepDataListMessage footstepDataList = HumanoidMessageTools.createFootstepDataListMessage(swingTime, transferTime);
-      footstepDataList.setExecutionTiming(ExecutionTiming.CONTROL_ABSOLUTE_TIMINGS);
+      footstepDataList.setExecutionTiming(ExecutionTiming.CONTROL_ABSOLUTE_TIMINGS.toByte());
 
       // step in place but do some fancy foot motion
       RobotSide robotSide = RobotSide.LEFT;
@@ -92,8 +92,8 @@ public abstract class AvatarFootstepDataMessageSwingTrajectoryTest implements Mu
       FrameQuaternion footOrientation = new FrameQuaternion(soleFrame);
 
       FootstepDataMessage footstep = new FootstepDataMessage();
-      footstep.setRobotSide(robotSide);
-      footstep.setTrajectoryType(TrajectoryType.WAYPOINTS);
+      footstep.setRobotSide(robotSide.toByte());
+      footstep.setTrajectoryType(TrajectoryType.WAYPOINTS.toByte());
       footstep.setTimings(swingTime, initialTransferTime);
 
       double radius = robotScale * 0.10;
