+++ ihmc-avatar-interfaces/src/test/java/us/ihmc/avatar/roughTerrainWalking/HumanoidLineContactWalkingTest.java
@@ -116,7 +116,7 @@ public abstract class HumanoidLineContactWalkingTest implements MultiRobotTestIn
 
          footstepData.setLocation(placeToStepInWorld);
          footstepData.setOrientation(new Quaternion(0.0, 0.0, 0.0, 1.0));
-         footstepData.setRobotSide(robotSide);
+         footstepData.setRobotSide(robotSide.toByte());
          message.add(footstepData);
 
          drcSimulationTestHelper.send(message);
@@ -162,7 +162,7 @@ public abstract class HumanoidLineContactWalkingTest implements MultiRobotTestIn
 
          footstepData.setLocation(placeToStepInWorld);
          footstepData.setOrientation(new Quaternion(0.0, 0.0, 0.0, 1.0));
-         footstepData.setRobotSide(robotSide);
+         footstepData.setRobotSide(robotSide.toByte());
          message.add(footstepData);
 
          drcSimulationTestHelper.send(message);
@@ -210,7 +210,7 @@ public abstract class HumanoidLineContactWalkingTest implements MultiRobotTestIn
 
          footstepData.setLocation(placeToStepInWorld);
          footstepData.setOrientation(new Quaternion(0.0, 0.0, 0.0, 1.0));
-         footstepData.setRobotSide(robotSide);
+         footstepData.setRobotSide(robotSide.toByte());
          message.add(footstepData);
 
          drcSimulationTestHelper.send(message);
@@ -382,7 +382,7 @@ public abstract class HumanoidLineContactWalkingTest implements MultiRobotTestIn
       for (RobotSide robotSide : RobotSide.values)
       {
          ArmTrajectoryMessage armTrajectoryMessage = new ArmTrajectoryMessage();
-         armTrajectoryMessage.robotSide = robotSide;
+         armTrajectoryMessage.robotSide = robotSide.toByte();
          double[] armConfig = straightArmConfigs.get(robotSide);
          armTrajectoryMessage.jointspaceTrajectory.jointTrajectoryMessages = new OneDoFJointTrajectoryMessage[armConfig.length];
          for (int i = 0; i < armConfig.length; i++)
